Moved LoggerWithSubscription method from utils to backendutils (#15582)
* Moved LoggerWithSubscription method from utils to backendutils

* image bump
