+++ components/eventing-controller/pkg/backend/nats/jetstream/jetstream.go
@@ -10,6 +10,8 @@ import (
 	"sync"
 	"time"
 
+	backendutils "github.com/kyma-project/kyma/components/eventing-controller/pkg/backend/utils"
+
 	cev2 "github.com/cloudevents/sdk-go/v2"
 	cev2protocol "github.com/cloudevents/sdk-go/v2/protocol"
 	"github.com/nats-io/nats.go"
@@ -149,7 +151,7 @@ func (js *JetStream) Initialize(connCloseHandler backendnats.ConnClosedHandler) 
 }
 
 func (js *JetStream) SyncSubscription(subscription *eventingv1alpha1.Subscription) error {
-	log := utils.LoggerWithSubscription(js.namedLogger(), subscription)
+	log := backendutils.LoggerWithSubscription(js.namedLogger(), subscription)
 	subKeyPrefix := backendnats.CreateKeyPrefix(subscription)
 	if err := js.checkJetStreamConnection(); err != nil {
 		return err
@@ -183,7 +185,7 @@ func (js *JetStream) SyncSubscription(subscription *eventingv1alpha1.Subscriptio
 }
 
 func (js *JetStream) DeleteSubscription(subscription *eventingv1alpha1.Subscription) error {
-	log := utils.LoggerWithSubscription(js.namedLogger(), subscription)
+	log := backendutils.LoggerWithSubscription(js.namedLogger(), subscription)
 
 	// loop over the global list of subscriptions
 	// and delete any related JetStream subscription
