+++ resources/eventing/values.yaml
@@ -5,7 +5,7 @@ global:
   images:
     eventing_controller:
       name: eventing-controller
-      version: PR-15556
+      version: PR-15582
       pullPolicy: "IfNotPresent"
     publisher_proxy:
       name: event-publisher-proxy
